2.12.0 feature release

+++ play.it-2/games/play-beatbuddy.sh
@@ -1,8 +1,9 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
 # Copyright (c) 2015-2020, Antoine "vv221/vv222" Le Gonidec
+# Copyright (c) 2016-2020, Mopi
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -30,41 +31,40 @@ set -o errexit
 
 ###
 # Beatbuddy: Tale of the Guardians
-# build native Linux packages from the original installers
-# send your bug reports to vv221@dotslashplay.it
+# build native packages from the original installers
+# send your bug reports to contact@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20200608.2
 
 # Set game-specific variables
 
 GAME_ID='beatbuddy'
 GAME_NAME='Beatbuddy: Tale of the Guardians'
 
-ARCHIVES_LIST='ARCHIVE_HUMBLE'
+ARCHIVES_LIST='
+ARCHIVE_HUMBLE_0
+'
 
-ARCHIVE_HUMBLE='BeatbuddyLinux1439603370.zip'
-ARCHIVE_HUMBLE_URL='https://www.humblebundle.com/store/beatbuddy'
-ARCHIVE_HUMBLE_MD5='156d19b327a02ac4a277f6f6ad4e188e'
-ARCHIVE_HUMBLE_SIZE='1100000'
-ARCHIVE_HUMBLE_VERSION='1.0-humble150815'
+ARCHIVE_HUMBLE_0='BeatbuddyLinux1439603370.zip'
+ARCHIVE_HUMBLE_0_MD5='156d19b327a02ac4a277f6f6ad4e188e'
+ARCHIVE_HUMBLE_0_URL='https://www.humblebundle.com/store/beatbuddy'
+ARCHIVE_HUMBLE_0_SIZE='1100000'
+ARCHIVE_HUMBLE_0_VERSION='1.0-humble150815'
 
 ARCHIVE_GAME_BIN_PATH='Beatbuddy'
-ARCHIVE_GAME_BIN_FILES='./Beatbuddy.x86 ./Beatbuddy_Data/Mono ./Beatbuddy_Data/Plugins'
+ARCHIVE_GAME_BIN_FILES='Beatbuddy.x86 Beatbuddy_Data/Mono Beatbuddy_Data/Plugins'
 
 ARCHIVE_GAME_DATA_PATH='Beatbuddy'
-ARCHIVE_GAME_DATA_FILES='./Beatbuddy_Data'
+ARCHIVE_GAME_DATA_FILES='Beatbuddy_Data'
 
 DATA_DIRS='./logs'
-DATA_FILES='./Beatbuddy.x86'
 
 APP_MAIN_TYPE='native'
 APP_MAIN_EXE='Beatbuddy.x86'
 # shellcheck disable=SC2016
 APP_MAIN_OPTIONS='-logFile ./logs/$(date +%F-%R).log'
-APP_MAIN_ICONS_LIST='APP_MAIN_ICON'
 APP_MAIN_ICON='Beatbuddy_Data/Resources/UnityPlayer.png'
-APP_MAIN_ICON_RES='128'
 
 PACKAGES_LIST='PKG_DATA PKG_BIN'
 
@@ -72,50 +72,59 @@ PKG_DATA_ID="${GAME_ID}-data"
 PKG_DATA_DESCRIPTIOn='data'
 
 PKG_BIN_ARCH='32'
-PKG_BIN_DEPS_DEB="$PKG_DATA_ID, libglu1-mesa | libglu1, libxcursor1, libasound2-plugins"
-PKG_BIN_DEPS_ARCH="$PKG_DATA_ID lib32-glu lib32-libxcursor lib32-alsa-plugins"
+PKG_BIN_DEPS="$PKG_DATA_ID glibc libstdc++ glu glx xcursor gtk2 alsa"
+PKG_BIN_DEPS_ARCH='lib32-libx11'
+PKG_BIN_DEPS_DEB='libx11-6'
+PKG_BIN_DEPS_GENTOO='x11-libs/libX11[abi_x86_32]'
 
 # Load common functions
-target_version='2.3'
+
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
 set_standard_permissions "$PLAYIT_WORKDIR/gamedata"
+prepare_package_layout
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
-PKG='PKG_BIN'
-organize_data 'GAME_BIN' "$PATH_GAME"
+# Get game icon
 
 PKG='PKG_DATA'
-organize_data 'GAME_DATA' "$PATH_GAME"
-
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+icons_get_from_package 'APP_MAIN'
 
 # Write launchers
 
 PKG='PKG_BIN'
-write_launcher 'APP_MAIN'
+launchers_write 'APP_MAIN'
 
 # Build package
 
-postinst_icons_linking 'APP_MAIN'
-write_metadata 'PKG_DATA'
-write_metadata 'PKG_BIN'
+write_metadata
 build_pkg
 
 # Clean up
